Removed all old pages since we don't need them anymore.
Changed the way we search for flights to make it easier to read and
handle checks and multiple destinations. 

Customers can now search for all flights as long as they exist. An error
message should appear if the flight is not available. 

Files: find_flights.jsp and display_flights.jsp are no longer used but I
am keeping them for now just in case I need to check something.